[parsers/schema, pipeline/config] Migrate parsing logic and remote sources
