+++ ocebuild/parsers/__init__.py
@@ -9,5 +9,6 @@ from ocebuild.parsers.asl import *
 from ocebuild.parsers.dict import *
 from ocebuild.parsers.plist import *
 from ocebuild.parsers.regex import *
+from ocebuild.parsers.schema import *
 from ocebuild.parsers.types import *
 from ocebuild.parsers.yaml import *
